Katmanlı Mimarinin Geliştirilmesi Wcf servisler katmanında ProductService oluşturuldu Interfaces katmanımızdaki IProductService'den miras aldı buradaki metodları OperationContract olarak tanımladık yani bunların bir servis operasyonları olduğunu belirttik. WcfLibrary kütüphanemizdeki servislerimizi yayınlamak için WcfllShost kütüphanesindeki web config'i düzenleyip servislerimizin yolunu ve erişim adreslerini tanımladık. WebUI katmanında ProductContoreller oluşturuldu Urlimizden Product requesti yapıldığında ürünlerin görüntülenmesini sağldık. Dal>ProductManager ile çektik NorhwindDbContext

+++ AspNetMvcNorthwind.Entities/Product.cs
@@ -10,7 +10,7 @@ namespace AspNetMvcNorthwind.Entities
     {
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public int UnitsInStock { get; set; }
+        public Int16 UnitsInStock { get; set; }
         public decimal UnitPrice { get; set; }
         public int CategoryID { get; set; }
     }
